update procs and logic to include compliance unit info

+++ CETrackerApi/CETracker.DbUp/Stored Procedures/ce.ExperienceAmount_U_I.sql
@@ -8,7 +8,7 @@ GO
 CREATE PROCEDURE ce.ExperienceAmount_U_I
 	@ExperienceId INT
 	,@UnitId INT
-	,@Amount INT
+	,@Amount DECIMAL(4,1)
 	,@UpdateUserId INT
 AS
